train: move the run out of main() into train_job.h
Everything a run decides -- resume validation, pre-encode, graph rebuilds, the
epoch loop, checkpointing -- lived in sa3-train's main(), so libsa3 could not
reach any of it. run_training() takes hooks for logging, per-step metrics,
cancellation, and host-supplied audio; sa3-train is now argv and printing.

Same seed and dataset produce a byte-identical adapter-final.gguf. 39/39 ctest.
